More Updates!
Have not included Gs latest map changes yet, but I have added more
untested mission system things. Added in a system to randomly select a
mission after a 10min timer. Also added in a global message sent to all
players so people know a mission is going to start. Smoke has been added
for where the chopper drops the payload.. Still need to add some more
mission markers as the missions will select a random position to fly to
and drop the box. More on that later though.